refactor(storyreader): create story loading module for text and live2d reader
