+++ src/components/story-selector/StorySelector.tsx
@@ -2,30 +2,9 @@ import React, { useMemo } from "react";
 import { useTranslation } from "react-i18next";
 import { Route, Switch, useRouteMatch } from "react-router-dom";
 
-import type {
-  ServerRegion,
-  IUnitProfile,
-  IUnitStory,
-  IEventStory,
-  IEventInfo,
-  ICharaProfile,
-  ICardEpisode,
-  ICardInfo,
-  IArea,
-  ISpecialStory,
-} from "../../types.d";
+import type { ServerRegion } from "../../types.d";
 
-import { useCachedData } from "../../utils";
-import { useAssetI18n, useCharaName } from "../../utils/i18n";
-import {
-  Breadcrumbs,
-  Grid,
-  Typography,
-  Card,
-  CardContent,
-  styled,
-} from "@mui/material";
-import LinkNoDecoration from "../../components/styled/LinkNoDecoration";
+import { Grid, Typography, Card, CardContent, styled } from "@mui/material";
 import LinkNoDecorationAlsoNoHover from "../../components/styled/LinkNoDecorationAlsoHover";
 const CardSelect = styled(Card)`
   &:hover {
@@ -34,6 +13,7 @@ const CardSelect = styled(Card)`
   }
 `;
 
+import Path from "./Path";
 import EventStory from "./EventStory";
 import UnitStory from "./UnitStory";
 import CharaStory from "./CharaStory";
@@ -49,19 +29,7 @@ const StorySelector: React.FC<{
   }) => void;
 }> = ({ onSetStory }) => {
   const { t } = useTranslation();
-  const { getTranslated } = useAssetI18n();
-  const getCharaName = useCharaName();
   const { path } = useRouteMatch();
-
-  const [unitProfiles] = useCachedData<IUnitProfile>("unitProfiles");
-  const [unitStories] = useCachedData<IUnitStory>("unitStories");
-  const [eventStories] = useCachedData<IEventStory>("eventStories");
-  const [events] = useCachedData<IEventInfo>("events");
-  const [characterProfiles] = useCachedData<ICharaProfile>("characterProfiles");
-  const [cardEpisodes] = useCachedData<ICardEpisode>("cardEpisodes");
-  const [cards] = useCachedData<ICardInfo>("cards");
-  const [areas] = useCachedData<IArea>("areas");
-  const [specialStories] = useCachedData<ISpecialStory>("specialStories");
 
   const handleSetStory = onSetStory;
 
@@ -114,198 +82,7 @@ const StorySelector: React.FC<{
 
   return (
     <>
-      <Route>
-        {({ location }) => {
-          const pathnames = location.pathname.split("/").filter((x) => x);
-          console.log(pathnames);
-          return (
-            <Breadcrumbs>
-              {pathnames.map((pathname, idx) => {
-                const last = idx === pathnames.length - 1;
-                const to = `/${pathnames.slice(0, idx + 1).join("/")}`;
-
-                let name = "";
-                if (idx === 0) {
-                  name = t("common:storyReader");
-                } else if (
-                  idx === 1 &&
-                  Object.keys(catagory).includes(pathname)
-                ) {
-                  name = catagory[pathname].breadcrumbName;
-                } else if (idx >= 2) {
-                  switch (pathnames[1]) {
-                    case "eventStory":
-                      if (events && idx === 2) {
-                        const found = events.find(
-                          (ev) => ev.id === Number(pathname)
-                        );
-                        if (found) {
-                          name = getTranslated(
-                            `event_name:${pathname}`,
-                            found.name
-                          );
-                        }
-                      }
-                      if (eventStories && idx === 3) {
-                        const found = eventStories.find(
-                          (es) => es.eventId === Number(pathnames[2])
-                        );
-                        if (found) {
-                          const episode = found.eventStoryEpisodes.find(
-                            (ese) => ese.episodeNo === Number(pathname)
-                          );
-                          if (episode) {
-                            name = getTranslated(
-                              `event_story_episode_title:${episode.eventStoryId}-${episode.episodeNo}`,
-                              episode.title
-                            );
-                          }
-                        }
-                      }
-                      break;
-                    case "unitStory":
-                      if (unitProfiles && idx === 2) {
-                        const found = unitProfiles.find(
-                          (unit) => unit.unit === pathname
-                        );
-                        if (found) {
-                          name = getTranslated(
-                            `unit_profile:${found.unit}.name`,
-                            found.unitName
-                          );
-                        }
-                      }
-                      if (unitStories) {
-                        const found = unitStories.find(
-                          (us) => us.unit === pathnames[2]
-                        );
-                        if (found && idx === 3) {
-                          const chapter = found.chapters.find(
-                            (cp) => cp.chapterNo === Number(pathname)
-                          );
-                          if (chapter) {
-                            name = getTranslated(
-                              `unit_story_chapter_title:${chapter.unit}-${chapter.chapterNo}`,
-                              chapter.title
-                            );
-                          }
-                        }
-                        if (found && idx === 4) {
-                          const chapter = found.chapters.find(
-                            (cp) => cp.chapterNo === Number(pathnames[3])
-                          );
-                          if (chapter) {
-                            const episode = chapter.episodes.find(
-                              (ep) => ep.episodeNo === Number(pathname)
-                            );
-                            if (episode) {
-                              name = getTranslated(
-                                `unit_story_episode_title:${episode.unit}-${episode.chapterNo}-${episode.episodeNo}`,
-                                episode.title
-                              );
-                            }
-                          }
-                        }
-                      }
-                      break;
-                    case "charaStory":
-                      if (characterProfiles && idx === 2) {
-                        const found = characterProfiles.find(
-                          (cp) => cp.characterId === Number(pathname)
-                        );
-                        if (found) {
-                          name = getCharaName(found.characterId) || "";
-                        }
-                      }
-                      break;
-                    case "cardStory":
-                      if (characterProfiles && idx === 2) {
-                        const found = characterProfiles.find(
-                          (cp) => cp.characterId === Number(pathname)
-                        );
-                        if (found) {
-                          name = getCharaName(found.characterId) || "";
-                        }
-                      }
-                      if (cards && idx === 3) {
-                        const card = cards.find(
-                          (card) => card.id === Number(pathname)
-                        );
-                        if (card) {
-                          name = getTranslated(
-                            `card_prefix:${card.id}`,
-                            card.prefix
-                          );
-                        }
-                      }
-                      if (cardEpisodes && idx === 4) {
-                        const episode = cardEpisodes.find(
-                          (cep) => cep.id === Number(pathname)
-                        );
-                        if (episode) {
-                          name = getTranslated(
-                            `card_episode_title:${episode.title}`,
-                            episode.title
-                          );
-                        }
-                      }
-                      break;
-                    case "areaTalk":
-                      if (areas && idx === 2) {
-                        const area = areas.find(
-                          (area) => area.id === Number(pathname)
-                        );
-                        if (area) {
-                          name = getTranslated(
-                            `area_name:${area.id}`,
-                            area.name
-                          );
-                        }
-                      }
-                      if (idx === 3) {
-                        name = pathname;
-                      }
-                      break;
-                    case "specialStory":
-                      if (specialStories) {
-                        if (idx === 2) {
-                          const chapter = specialStories.find(
-                            (sp) => sp.id === Number(pathname)
-                          );
-                          if (chapter) {
-                            name = chapter.title;
-                          }
-                        } else if (idx === 3) {
-                          const chapter = specialStories.find(
-                            (sp) => sp.id === Number(pathnames[2])
-                          );
-                          if (chapter) {
-                            const episode = chapter.episodes.find(
-                              (ep) => ep.episodeNo === Number(pathname)
-                            );
-                            if (episode) {
-                              name = episode.title;
-                            }
-                          }
-                        }
-                      }
-                  }
-                }
-
-                return last ? (
-                  <Typography color="textPrimary" key={to}>
-                    {name}
-                  </Typography>
-                ) : (
-                  <LinkNoDecoration key={to} to={to}>
-                    {name}
-                  </LinkNoDecoration>
-                );
-              })}
-            </Breadcrumbs>
-          );
-        }}
-      </Route>
+      <Path catagory={catagory} />
       <Switch>
         <Route path={`${path}`} exact>
           <Grid container spacing={1}>
